pyproject.toml


[build-system]
requires=["setuptools"]
build_backend="setuptools.build_meta"


[project]
name="pixelator"
dependencies=["Pillow"]
dynamic=["version"]

[project.scripts]
pixelator="pixelator.__main__:main"

[tools.setuptools.dynamic]
version = {attr = 'pixelator.__version__'}
